+++ core/local/src/main/java/com/anshtya/core/local/database/entity/TrendingContentEntity.kt
@@ -10,7 +10,7 @@ data class TrendingContentEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
     @ColumnInfo(name = "remote_id")
-    val remoteId: Long,
+    val remoteId: Int,
     @ColumnInfo(name = "image_path")
     val imagePath: String,
     val name: String
